DPR-482 move editor code into dedicated package
