+++ cli/src/main/kotlin/uk/gov/justice/digital/cli/editor/MultilineEditor.kt
@@ -1,10 +1,10 @@
-package uk.gov.justice.digital.cli.command
+package uk.gov.justice.digital.cli.editor
 
 import org.jline.keymap.BindingReader
 import org.jline.keymap.KeyMap
 import org.jline.terminal.Terminal
 import org.jline.utils.InfoCmp
-import uk.gov.justice.digital.cli.command.MultilineEditor.Operation.*
+import uk.gov.justice.digital.cli.editor.MultilineEditor.Operation.*
 import uk.gov.justice.digital.cli.session.ConsoleSession
 import kotlin.math.min
